style: Prevent dash from overlapping content
In order to center the view selector, the dash has been moved to a
separate layer, which means that it will overlap the app picker and
search results if the available width is smaller than the maximum
width that the content will request. Fix this by adding enough
horizontal padding to account for the width the dash will have at
its largest icon size.

https://bugzilla.gnome.org/show_bug.cgi?id=695471